completed local testing

+++ database/migrations/2023_01_15_153420_create_performances_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -17,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('round_id');
             $table->uuid('user_id');
-            $table->integer('duration');
+            $table->bigInteger('duration');
             $table->timestamps();
 
             $table->foreign('round_id')->references('id')->on('rounds');
